feat(tables): rename DefaultTable to DataTable
Also move SimpleTable & TableWithSelectableRows to `tables/` directory.
